Moved getLink onto StateHandler
Means that Fluent and State navigators can call in same way
