+++ Navigation/src/StateNavigator.ts
@@ -45,7 +45,7 @@ class StateNavigator {
         this.stateContext.crumbs = data[state.crumbTrailKey];
         delete data[state.crumbTrailKey];
         this.stateContext.data = data;
-        this.stateContext.nextCrumb = new Crumb(data, state, url, this.getLink(state, data), false);
+        this.stateContext.nextCrumb = new Crumb(data, state, url, this.stateHandler.getLink(state, data), false);
         this.stateContext.previousState = null;
         this.stateContext.previousData = {};
         if (this.stateContext.crumbs.length > 0) {
@@ -81,20 +81,7 @@ class StateNavigator {
         if (!this.states[stateKey])
             throw new Error(stateKey + ' is not a valid State');
         var { crumbs, nextCrumb } = this.stateContext;
-        return this.getLink(this.states[stateKey], navigationData, crumbs, nextCrumb);
-    }
-
-    private getLink(state: State, navigationData: any, crumbs?: Crumb[], nextCrumb?: Crumb): string {
-        var crumbTrail = [];
-        if (crumbs) {
-            crumbs = crumbs.slice();
-            if (nextCrumb)
-                crumbs.push(nextCrumb);
-            crumbs = state.truncateCrumbTrail(state, crumbs);
-            for(var i = 0; i < crumbs.length; i++)
-                crumbTrail.push(crumbs[i].crumblessUrl)
-        }
-        return this.stateHandler.getNavigationLink(state, navigationData, crumbTrail);
+        return this.stateHandler.getLink(this.states[stateKey], navigationData, crumbs, nextCrumb);
     }
 
     canNavigateBack(distance: number) {
@@ -121,7 +108,7 @@ class StateNavigator {
 
     getRefreshLink(navigationData?: any): string {
         var { crumbs, nextCrumb } = this.stateContext;
-        return this.getLink(this.stateContext.state, navigationData, crumbs, nextCrumb);
+        return this.stateHandler.getLink(this.stateContext.state, navigationData, crumbs, nextCrumb);
     }
 
     navigateLink(url: string, historyAction: 'add' | 'replace' | 'none' = 'add', history = false) {
@@ -167,7 +154,7 @@ class StateNavigator {
 
     fluent(withContext = false): FluentNavigator {
         var stateContext = !withContext ? null : this.stateContext;
-        return new FluentNavigator(this.states, this.getLink, this.stateHandler.parseNavigationLink, stateContext);
+        return new FluentNavigator(this.states, this.stateHandler, stateContext);
     }
     
     start(url?: string) {
